requirements.txt
requests
python-dotenv
pytest
playwright
